ERgraph_random_removal.m
Creates an ER graph of size 'N', with a probability 'p' such that a giant component exists initially. Then successively removes increasing fractions of nodes (in increments of 1/'division'), finds the giant component size after the removal and then restores the graph to the initial stage for the next removal. Repeats the process for 'trials' number of times. Prints the theoretical cut-off occupation probability, i.e. if a fraction of nodes less than the given probability remain, then the graph loses its giant component (its size falls significantly). Plots the relative size of the giant component with respect to the occupation probability.